--- app/components/Footer/Footer.js
@@ -1,11 +1,0 @@
-'use strict';
-
-import React from 'react';
-
-const Footer = () => (
-  <footer className="credits">
-    Made with <span className="heart"></span>by <a href="https://github.com/afonsopacifer" className="link" target="_blank">@afonsopacifer</a>
-  </footer>
-);
-
-export default Footer;
