--- app/index.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" manifest="pomodoro.appcache">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, minimum-scale=1.0, maximum-scale=1.0">
-    <link rel="icon" href="img/favicon.png" sizes="32x32">
-    <link href='https://fonts.googleapis.com/css?family=Roboto:400,300,900' rel='stylesheet' type='text/css'>
-
-    <!-- SEO -->
-    <link rel="manifest" href="manifest.json">
-
-    <!-- SEO -->
-		<meta name="description" content="Pomodoro timer for developers.">
-    <title>Pomodoro timer</title>
-
-    <!-- Open graph -->
-		<meta property="og:type" content="website">
-		<meta property="og:title" content="Pomodoro timer">
-		<meta property="og:description" content="Pomodoro timer for developers.">
-		<meta property="og:url" content="http://afonsopacifer.github.io/react-pomodoro/">
-		<meta property="og:image" content="http://afonsopacifer.github.io/react-pomodoro/img/share.png">
-
-		<!-- Twitter cards -->
-		<meta name="twitter:card" content="summary_large_image">
-		<meta name="twitter:title" content="Pomodoro timer">
-		<meta name="twitter:description" content="Pomodoro timer for developers.">
-		<meta name="twitter:url" content="http://afonsopacifer.github.io/react-pomodoro/">
-		<meta name="twitter:image" content="http://afonsopacifer.github.io/react-pomodoro/img/share.png">
-
-		<!-- CSS <3 -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/3.0.3/normalize.css">
-    <link rel="stylesheet" href="css/style.css">
-
-
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="text/JavaScript" src="bundle.js"></script>
-  </body>
-</html>
