--- app/js/app.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDom from 'react-dom';
-import Pomodoro from './../components/Pomodoro/Pomodoro';
-
-ReactDom.render(<Pomodoro />, document.getElementById('app'));
