--- app/manifest.json
@@ -1,13 +1,0 @@
-{
-  "name": "React Pomodoro",
-  "icons": [
-    {
-      "src": "img/launcher-icon-1x.png",
-      "sizes": "48x48",
-      "type": "image/png",
-      "density": 1.0
-    }
-  ],
-  "start_url": "index.html",
-  "display": "standalone"
-}
