+++ package.json
@@ -26,9 +26,14 @@
         "@testing-library/jest-dom": "^5.11.4",
         "@testing-library/react": "^11.1.0",
         "@testing-library/user-event": "^12.1.10",
+        "axios": "^0.21.1",
+        "bootstrap": "^4.6.0",
+        "font-awesome": "^4.7.0",
         "jquery": "^3.5.1",
+        "mousetrap": "^1.6.5",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
+        "react-router-dom": "^5.2.0",
         "react-scripts": "4.0.1",
         "web-vitals": "^0.2.4"
     },
