+++ src/pomodoro/Footer.js
@@ -1,0 +1,19 @@
+import React from "react";
+const Footer = () => (
+    <footer className="credits">
+        Made with{" "}
+        <i
+            style={{ color: "red" }}
+            className="fa fa-heart"
+            aria-hidden="true"
+        />{" "}
+        by Koders
+        {/* <a
+            href="https://github.com/afonsopacifer"
+            className="link"
+            target="_blank"> */}
+        {/* </a> */}
+    </footer>
+);
+
+export default Footer;
