+++ src/pomodoro/img/launcher-icon-1x.png
